enhanced ris_pull to also grab c:\Program Files\RISADemo, if available

+++ DevTest/Hyper-V/risa_pull.ps1
@@ -1,7 +1,19 @@
 md -Force "$env:USERPROFILE\Desktop\Pull"
 md -Force "$env:USERPROFILE\Desktop\Pull\Documents"
 md -Force "$env:USERPROFILE\Desktop\Pull\ProgramFiles"
+md -Force "$env:USERPROFILE\Desktop\Pull\ProgramFilesDemo"
 md -Force "$env:USERPROFILE\Desktop\Pull\ProgramData"
-Copy-Item -Path "$env:USERPROFILE\Documents\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\Documents" -PassThru -Recurse
-Copy-Item -Path "C:\Program Files\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\ProgramFiles" -PassThru -Recurse
-Copy-Item -Path "C:\ProgramData\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\ProgramData" -PassThru -Recurse
+
+
+if (Test-Path -Path "$env:USERPROFILE\Desktop\Pull\Documents") {
+    Copy-Item -Path "$env:USERPROFILE\Documents\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\Documents" -PassThru -Recurse
+}
+if (Test-Path -Path "C:\Program Files\RISA") {
+    Copy-Item -Path "C:\Program Files\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\ProgramFiles" -PassThru -Recurse
+}
+if (Test-Path -Path "C:\Program Files\RISADemo") {
+    Copy-Item -Path "C:\Program Files\RISADemo\*" -Destination "$env:USERPROFILE\Desktop\Pull\ProgramFilesDemo" -PassThru -Recurse
+}
+if (Test-Path -Path "C:\ProgramData\RISA") {
+    Copy-Item -Path "C:\ProgramData\RISA\*" -Destination "$env:USERPROFILE\Desktop\Pull\ProgramData" -PassThru -Recurse
+}
